refactor(example): refactored example
This commit adds refactoring, the code that will discover and handle all the services and
characteristics of interest offered by the connected peripheral and print their uuid.  Ultimately
the app will be interested in only one services: the heart rate monitory service. This services is
identified by their UUID -- the app will print an extra message when it discovers these services of
interest.  The characteristic used to report heart rate measurements is one that the app can
subscribe to, so we do that. This will automatically result in receiving
didUpdateValueForCharacteristic events whenever the heart rate monitor reports a new heart rate
value.